requirements.txt
streamlit>=1.32.0
youtube-transcript-api>=0.6.2
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
